requirements.txt
django==2.2.7
djangorestframework==3.10.3
pubnub>=4.1.7
gunicorn
requests